Move sign out button to navbar
